pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sheaf-serve"
version = "0.5.0"
description = "Unified serving layer for non-text foundation models"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Alex Korbonits", email = "alexkorbonits@gmail.com" }]
keywords = ["serving", "inference", "foundation-models", "ray", "mlops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "ray[serve]>=2.10.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
time-series = [
    "chronos-forecasting>=1.0.0",
    "timesfm[torch]>=1.0.0",
]
tabular = [
    "tabpfn>=2.0.0",
]
moirai = [
    "uni2ts>=2.0.0",
]
audio = [
    "openai-whisper>=20240930",
    "faster-whisper>=1.2.1",
]
audio-generation = [
    "transformers>=4.37.0",
    "torch>=2.0.0",
]
tts = [
    "transformers>=4.31.0",
    "torch>=2.0.0",
]
kokoro = [
    "kokoro>=0.9.2",
    "soundfile>=0.12.1",
]
vision = [
    "open-clip-torch>=3.0.0",
    "transformers>=4.37.0",
    "sam2>=1.0",
]
pose = [
    "transformers>=4.46.0",
    "torch>=2.0.0",
]
optical-flow = [
    "torchvision>=0.15.0",
    "torch>=2.0.0",
]
multimodal-generation = [
    "diffusers>=0.26.0",
    "torch>=2.0.0",
    "Pillow>=9.0.0",
]
lidar = [
    "torch>=2.0.0",
]
molecular = [
    # ESM-3 requires Python 3.12+; marker prevents resolution errors on 3.11
    "esm>=3.0.0 ; python_full_version >= '3.12'",
]
weather = [
    # graphcast 1.0.0 requires Python 3.11 only; pin <1.0.0 for 3.11 + 3.12 support
    "graphcast>=0.1.0,<1.0.0",
    "dm-haiku>=0.0.12",
    "jax>=0.4.25",
    "xarray>=2024.1.0",
]
earth-observation = [
    "transformers>=4.37.0",  # Prithvi uses trust_remote_code=True
    "torch>=2.0.0",
]
genomics = [
    "transformers>=4.37.0",  # Nucleotide Transformer via AutoModel + AutoTokenizer
    "torch>=2.0.0",
]
materials = [
    "mace-torch>=0.3.0",  # MACE-MP-0 universal interatomic potential; brings torch + ase
]
small-molecule = [
    "transformers>=4.37.0",  # MolFormer via AutoModel + AutoTokenizer (trust_remote_code)
    "torch>=2.0.0",
]
multimodal = [
    # imagebind itself is not on PyPI — install from the official Meta repo:
    #   pip install git+https://github.com/facebookresearch/ImageBind.git
    # This extra provides the remaining dependencies (PyTorch stack).
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
]
diffusion = [
    "diffusers>=0.30.0",
    "torch>=2.0.0",
    "transformers>=4.37.0",  # diffusers depends on transformers for text encoders
]
video = [
    "transformers>=4.40.0",  # VideoMAE AutoModel support
    "torch>=2.0.0",
]
feast = [
    "feast>=0.40.0",
]
metrics = [
    "prometheus-client>=0.20.0",
]
tracing = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]
modal = [
    "modal>=0.60.0",
    "cloudpickle>=3.0.0",
]
milvus = [
    "pymilvus>=2.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "cloudpickle>=3.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "ty>=0.0.0",
    "pre-commit>=3.7.0",
]
all = [
    # Excludes extras with narrow Python constraints or cross-extra conflicts:
    #   moirai:    uni2ts pins hydra-core==1.3; sam2 (vision) needs >=1.3.2
    #   molecular: esm requires Python 3.12+
    #   weather:   graphcast requires Python 3.11 only
    #   multimodal: imagebind not on PyPI; install from source separately
    "sheaf-serve[time-series,tabular,audio,audio-generation,tts,vision,genomics,small-molecule,materials,earth-observation,feast,milvus]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sheaf"]

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/sheaf"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
# uni2ts (moirai) pins hydra-core==1.3; sam2 (vision) requires hydra-core>=1.3.2.
# [all] includes vision, so it also conflicts with moirai.
conflicts = [
    [{ extra = "moirai" }, { extra = "vision" }],
    [{ extra = "moirai" }, { extra = "all" }],
]

[tool.uv.sources]
chronos-forecasting = { git = "https://github.com/amazon-science/chronos-forecasting.git" }

[dependency-groups]
dev = [
    "modal>=1.4.1",
]
